add pager objects

+++ src/Arrays/ArrayPage.php
@@ -38,7 +38,7 @@ final class ArrayPage implements Page
      */
     public function getCurrentPage()
     {
-        return floor($this->offset / $this->limit) + 1;
+        return (int) (floor($this->offset / $this->limit) + 1);
     }
 
     /**
